chore(wmg): move Filters into WMG2 folder
